Explorer Alpha 0.5
First functioning version of the bot. Takes any coordinates and automatically boats towards it.

Known issues:
Entering chat, pressing W, or basically any GUI causes the bot to stop boating forwards, you'll have to start holding W automatically
There is no way to quickly stop the bot for any reason. This is upcoming.
No pathfinding. It'll run into walls if it isnt moderated or you aren't on a known path
Throws errors when not in a boat. This will be fixed upon Main release.

Other then that, it's a good auto-pather. As long as you know the path and just want to sit back and wait to arrive, this is your best option so far.